refactor: js files and modules
